remove and rename files no longer necessary after refactor
