--- citrine/connection_tools/__init__.py
@@ -1,1 +1,0 @@
-#from citrine.connection_tools.connection import CitrineConnection, ManagedConnection
